requirements.txt
bcrypt
requests
python-dotenv
google-generativeai
matplotlib
pytest